requirements.txt
PyYAML==3.10
Twisted==12.2.0
beautifulsoup4==4.1.3
coverage==3.5.3
cyclone==1.0-rc12
distribute==0.6.28
iso8601==0.1.4
logilab-astng==0.24.1
logilab-common==0.58.1
mock==1.0.0
nose==1.2.1
oauth2==1.5.211
pyOpenSSL==0.13
pylint==0.26.0
pyodbc==3.0.6
python-faker==0.2.4
pyutil==1.9.4
requests==0.14.1
trialcoverage==0.3.12
txStatsD==1.0.0
wsgiref==0.1.2
zbase32==1.1.5
zope.interface==4.0.1
